% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt.R
\name{fmt}
\alias{fmt}
\title{fmt}
\usage{
fmt(x, decimales)
}
\arguments{
\item{x}{un nombre}

\item{decimales}{un entier, le nombre de decimales}
}
\value{
Le nombre formate.
}
\description{
Formate un nombre
}
\details{
La fonction formate avec
le nombre de decimales specifie, 
la virgule pour distinguer unites et decimales, 
l'espace comme separateur de milliers.
}
\section{Emploie }{

La fonction \url{https://www.rdocumentation.org/packages/base/versions/3.6.2/topics/format}
}

\section{Warning }{

N'utilise pas de notation scientifique
}

\examples{
fmt2Decimales(100, 2)
fmt2Decimales(1234.56789, 2)
}
\references{
\url{https://www.rdocumentation.org/packages/base/versions/3.6.2/topics/format} la fonction format
}
\seealso{
\url{https://www.rdocumentation.org/packages/base/versions/3.6.2/topics/print} pour afficher dans la console

Other fonctions pour formater: 
\code{\link{fmt1Decimale}()}
}
\author{
M. Martin, \email{martin@insee.fr}
}
\concept{fonctions pour formater}
